# backend/src/main/java/com/sai/backend/entity/Department.java
package com.sai.backend.entity;

import javax.persistence.*;
import javax.validation.constraints.Size;

@Entity
public class Department {

    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Integer departmentId;

    @Column(name = "department_name")
    @Size(max = 30)
    private String departmentName;

    @Column(name = "manager_id")
    private Integer managerId;


    protected Department() {

    }

    public Department(final String departmentName, final Integer managerId) {
        this.departmentName = departmentName;
        this.managerId = managerId;
    }

    public Integer getDepartmentId() {
        return departmentId;
    }

    public void setDepartmentId(Integer departmentId) {
        this.departmentId = departmentId;
    }

    public String getDepartmentName() {
        return departmentName;
    }

    public void setDepartmentName(String departmentName) {
        this.departmentName = departmentName;
    }

    public Integer getManagerId() {
        return managerId;
    }

    public void setManagerId(Integer managerId) {
        this.managerId = managerId;
    }
}